♻️refactor: 폴더명 명사형으로 수정
