+++ src/services/commit-analysis/analyze-commits.ts
@@ -1,6 +1,6 @@
 import { Commit } from "@/app/types/commit";
 import { TOKEN_LIMITS, REQUEST_INPUT_INTRO_MESSAGE } from "@/constants/open-ai";
-import chunkCommitsByTokens from "@/services/analyze-commits/helpers/chunk-commits-by-tokens";
+import chunkCommitsByTokens from "@/services/commit-analysis/helpers/chunk-commits-by-tokens";
 import {
   commitAnalysesSchema,
   overallAnalysisSchema,
