+++ src/services/commit-analysis/helpers/sort-merged-results.ts
